+++ tests/Gedmo/Mapping/Fixture/Document/Revisionable.php
@@ -1,0 +1,54 @@
+<?php
+
+/*
+ * This file is part of the Doctrine Behavioral Extensions package.
+ * (c) Gediminas Morkevicius <gediminas.morkevicius@gmail.com> http://www.gediminasm.org
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Gedmo\Tests\Mapping\Fixture\Document;
+
+use Doctrine\ODM\MongoDB\Mapping\Annotations as ODM;
+use Doctrine\ODM\MongoDB\Types\Type;
+use Gedmo\Mapping\Annotation as Gedmo;
+
+/**
+ * @ODM\Document(collection="revisionables")
+ *
+ * @Gedmo\Revisionable
+ */
+#[ODM\Document(collection: 'revisionables')]
+#[Gedmo\Revisionable]
+class Revisionable
+{
+    /**
+     * @ODM\Id
+     */
+    #[ODM\Id]
+    private ?string $id = null;
+
+    /**
+     * @ODM\Field(type="string")
+     *
+     * @Gedmo\Versioned
+     */
+    #[ODM\Field(type: Type::STRING)]
+    #[Gedmo\Versioned]
+    private ?string $title = null;
+
+    public function getId(): ?string
+    {
+        return $this->id;
+    }
+
+    public function setTitle(string $title): void
+    {
+        $this->title = $title;
+    }
+
+    public function getTitle(): ?string
+    {
+        return $this->title;
+    }
+}